fix: jumping cursor and row not deleting after save
Before this commit, the items were set everytime content
changed on the froala editors, which caused the cursor to
jump (due to re-render). This commit fixes it by only setting
the widget content when text is changed in the editor and not
setting the items state. The items state is kept in sync with
widget's content when the editMode changes.

To fix the last items not deleting after save, we now set the
items state correctly, which in turn triggers widget.setContent effect.

Signed-off-by: Gaurav Bhorkar <gaurav.bhorkar@happeo.com>